Move ColorMarkupLabel
