+++ core/src/com/unciv/ui/components/widgets/ColorMarkupLabel.kt
@@ -1,6 +1,6 @@
 @file:Suppress("unused")
 
-package com.unciv.ui.components
+package com.unciv.ui.components.widgets
 
 import com.badlogic.gdx.graphics.Color
 import com.badlogic.gdx.graphics.Colors
@@ -8,6 +8,7 @@ import com.badlogic.gdx.graphics.g2d.GlyphLayout
 import com.badlogic.gdx.scenes.scene2d.ui.Label
 import com.unciv.Constants
 import com.unciv.models.translations.tr
+import com.unciv.ui.components.Fonts
 import com.unciv.ui.screens.basescreen.BaseScreen
 
 /** A Label allowing Gdx markup
